+++ src/__test__/use-async-data.test.ts
@@ -1,6 +1,8 @@
 import { describe, expect, test, vi } from 'vitest'
 import { useAsyncData } from '../use-async-data'
 
+const wait = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
+
 describe('useAsyncData', () => {
   test('非法参数', () => {
     // @ts-expect-error
@@ -51,5 +53,111 @@ describe('useAsyncData', () => {
     expect(querySum()).toBe(2)
     const { queryAsyncSum } = useAsyncData('asyncSum', () => new Promise((resolve) => setTimeout(() => resolve(sum(1, 1)), 100)))
     expect(queryAsyncSum()).resolves.toBe(2)
+  })
+
+  test('多次异步，依次起止', async () => {
+    const sumFunc = async (a: number, b: number) => {
+      await(a + b)
+      return a + b
+    }
+    const { querySum, sum } = useAsyncData('sum', sumFunc)
+    
+    const p1 = querySum(40, 30)
+    expect(sum.value).toBeUndefined()
+    await p1
+    expect(sum.value).toBe(70)
+
+    const p2 = querySum(50, 40)
+    // 上次计算结果
+    expect(sum.value).toBe(70)
+    await p2
+    expect(sum.value).toBe(90)
+  })
+
+  test('多次异步，结果顺序', async () => {
+    const sumFunc = async (a: number, b: number) => {
+      await wait(a + b)
+      return a + b
+    }
+    const { querySum, sum } = useAsyncData('sum', sumFunc)
+    
+    const p1 = querySum(40, 30)
+    const p2 = querySum(50, 40)
+
+    expect(sum.value).toBeUndefined()
+    await expect(p1).resolves.toBe(70)
+    expect(sum.value).toBe(70)
+    await expect(p2).resolves.toBe(90)
+    expect(sum.value).toBe(90)
+  })
+
+  test('多次异步，结果异序', async () => {
+    const sumFunc = async (a: number, b: number) => {
+      await wait(a + b)
+      return a + b
+    }
+    const { querySum, sum, sumExpired } = useAsyncData('sum', sumFunc)
+    
+    const p1 = querySum(170, 30)
+    const p2 = querySum(70, 30)
+
+    expect(sum.value).toBeUndefined()
+    await expect(p2).resolves.toBe(100)
+    expect(sum.value).toBe(100)
+
+    await expect(p1).resolves.toBe(200)
+    expect(sum.value).toBe(100)
+  })
+
+  test('调用异常', async () => {
+    const testFunc = (result: number, error?: Error) => {
+      if (error) throw error
+      return result
+    }
+    const { queryTest, test, testExpired } = useAsyncData('test', testFunc)
+
+    expect(test.value).toBeUndefined()
+    expect(queryTest(25)).toBe(test.value)
+
+    const error = new Error()
+    expect(() => queryTest(30, error)).toThrow(error)
+    expect(test.value).toBe(25)
+    expect(testExpired.value).toBeTruthy()
+  })
+
+  test('调用异步异常', async () => {
+    const testFunc = async (result: number, error?: Error, ms?: number) => {
+      if (ms) await wait(ms)
+      if (error) throw error
+      return result
+    }
+
+    const { queryTest, test, testExpired } = useAsyncData('test', testFunc)
+
+    expect(test.value).toBeUndefined()
+    const v = await queryTest(35, undefined, 30)
+    expect(v).toBe(test.value)
+
+    // 后者先到，展示后者数据，前者报错不影响
+    const p1 = queryTest(36, new Error(), 100)
+    const p2 = queryTest(51, undefined, 50)
+    const r2 = await p2
+    expect(r2).toBe(test.value)
+    expect(testExpired.value).toBeFalsy()
+    await expect(p1).rejects.toThrowError()
+    expect(test.value).toBe(51)
+    expect(testExpired.value).toBeFalsy()
+
+    // 后者先到，报错，前者后到，忽略后到的前者数据
+    const p3 = queryTest(37, undefined, 100)
+    const p4 = queryTest(52, new Error(), 50)
+    expect(testExpired.value).toBeFalsy()
+    await expect(p4).rejects.toThrowError()
+    expect(test.value).toBe(51)
+    expect(testExpired.value).toBeTruthy()
+
+    await expect(p3).resolves.toBe(37)
+    expect(test.value).toBe(37)
+    expect(testExpired.value).toBeTruthy()
   })
 })